requirements.txt
discord
audioop-lts
python-dotenv

pyyaml

pymongo
motor
aiosqlite